move tests to load_data_and_get_combined_routes
